streamlit_meditation_rec.py
import os
import logging
import streamlit as st
from dotenv import load_dotenv
from langchain_community.chat_message_histories import StreamlitChatMessageHistory
from langchain_core.messages import AIMessage, HumanMessage

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Load environment variables
load_dotenv()

# Initialize message history
msgs = StreamlitChatMessageHistory(key="langchain_messages")

# Streamlit page configuration
st.set_page_config(page_title="Simple Chatbot", page_icon="💬")
st.title("Simple Chatbot 💬")

# Placeholder for the LLM API key (replace with actual implementation)
api_token = os.getenv("TOGETHER_API_KEY")

# ---- LLM Initialization ----
def initialize_llm():
    """
    Initialize and return the LLM.
    Replace with your specific LLM initialization code.
    """
    # Placeholder: Replace with your LLM initialization logic
    return None

chat_llm = initialize_llm()

# ---- Reply Generation ----
def generate_reply(messages):
    """
    Generate a reply from the LLM based on the chat history.
    """
    # Placeholder: Implement the logic to use the LLM for generating a reply
    response = "This is a placeholder response. Implement your LLM logic here."
    return response.strip()

# Display existing chat messages
for msg in msgs.messages:
    with st.chat_message(msg.type):
        st.write(msg.content)

# Handle new user input
if prompt := st.chat_input("Type your message here..."):
    # Add user message to the history and display it
    user_message = HumanMessage(content=prompt)
    msgs.add_message(user_message)
    with st.chat_message("user"):
        st.write(prompt)

    # Generate AI response
    response = generate_reply(msgs.messages)

    # Add AI message to the history and display it
    ai_message = AIMessage(content=response)
    msgs.add_message(ai_message)
    with st.chat_message("ai"):
        st.write(response)

# Option to reset the chat
if st.sidebar.button("Reset Chat"):
    msgs.clear()
    st.experimental_rerun()
